refactor: Introduce CoreModule for services and HTTP interceptors
- Create a new CoreModule which contains services (RecipeService and ShoppingListService) and an HTTP interceptor (AuthInterceptor).
- Remove the providers (services and interceptors) from app.module.ts and add CoreModule to its imports.
- Import CommonModule in CoreModule.

The addition of CoreModule streamlines the organization of services and HTTP interceptors within the application. By grouping these core functionalities in a separate module, it declutters the AppModule and adheres to the Angular best practice of having a CoreModule handle singleton services and core functionalities.

This refactor aims to make the application more maintainable and structured, by ensuring that AppModule remains lean and only responsible for defining the root module, while CoreModule takes care of services and HTTP interceptors that are used application-wide.